refactor: unify income and income-category controllers
